+++ model/gurps/weapon_damage.go
@@ -477,7 +477,11 @@ func (w *WeaponDamage) resolveDiceSpec(s string) (d *dice.Dice, sub bool) {
 	if d, sub = parsePotentialDiceSpec(s); d != nil {
 		return d, sub
 	}
-	value := ResolveScript(w.Owner.Entity(), deferredNewScriptWeapon(w.Owner), s)
+	var entity *Entity
+	if w.Owner != nil {
+		entity = w.Owner.Entity()
+	}
+	value := ResolveScript(entity, deferredNewScriptWeapon(w.Owner), s)
 	value = strings.TrimPrefix(strings.TrimSpace(value), "+")
 	if sub = isDiceSubtraction(value); sub {
 		value = value[1:]