+++ ux/spell_table.go
@@ -30,6 +30,9 @@ func (p *spellListProvider) SpellList() []*gurps.Spell {
 }
 
 func (p *spellListProvider) SetSpellList(list []*gurps.Spell) {
+	for _, one := range list {
+		one.SetDataOwner(nil)
+	}
 	p.spells = list
 }